Rename current cross tests
